Prepare StealthLib 2.2.1 release
Fix installed-package and FetchContent consumers, PE forwarder resolution, IAT comparison, literal unlock lifetime, decoder edge cases, VM detection, and verification drift. Add release workflow that builds and runs CTest before publishing the tag release.

+++ stealthlib/detail/encryption.hpp
@@ -17,7 +17,7 @@
 
 namespace stealth::detail {
 
-template<size_t N, size_t Idx>
+template<size_t N, uint64_t Idx>
 struct encrypted_string_impl {
     char encrypted[N]{};
     char buffer[N + 1]{};
@@ -186,7 +186,7 @@ struct encrypted_string_impl {
 #endif
 };
 
-template<size_t N, size_t Idx>
+template<size_t N, uint64_t Idx>
 struct encrypted_wstring_impl {
     wchar_t encrypted[N]{};
     wchar_t buffer[N + 1]{};
